валидация роутов celebrate-ом вынесена в отдельный файл
