+++ app.js
@@ -52,36 +52,36 @@ app.use(express.json());
 
 // app.use(isAuthorized);
 
-const validateMovie = celebrate({
-  body: Joi.object().keys({
-    country: Joi.string().required(),
-    director: Joi.string().required(),
-    duration: Joi.number().required(),
-    year: Joi.number().required(),
-    description: Joi.string().required(),
-    image: Joi.string().required().custom((value, helpers) => {
-      if (validator.isURL(value, { protocols: ['http', 'https', 'ftp'], require_tld: true, require_protocol: true })) {
-        return value;
-      }
-      return helpers.message('Ссылка введена некорректно');
-    }),
-    trailerLink: Joi.string().required().custom((value, helpers) => {
-      if (validator.isURL(value, { protocols: ['http', 'https', 'ftp'], require_tld: true, require_protocol: true })) {
-        return value;
-      }
-      return helpers.message('Ссылка введена некорректно');
-    }),
-    nameRU: Joi.string().required(),
-    nameEN: Joi.string().required(),
-    thumbnail: Joi.string().required().custom((value, helpers) => {
-      if (validator.isURL(value)) {
-        return value;
-      }
-      return helpers.message('Ссылка введена некорректно');
-    }),
-    movieId: Joi.number().required(),
-  }),
-});
+// const validateMovie = celebrate({
+//   body: Joi.object().keys({
+//     country: Joi.string().required(),
+//     director: Joi.string().required(),
+//     duration: Joi.number().required(),
+//     year: Joi.number().required(),
+//     description: Joi.string().required(),
+//     image: Joi.string().required().custom((value, helpers) => {
+//       if (validator.isURL(value, { protocols: ['http', 'https', 'ftp'], require_tld: true, require_protocol: true })) {
+//         return value;
+//       }
+//       return helpers.message('Ссылка введена некорректно');
+//     }),
+//     trailerLink: Joi.string().required().custom((value, helpers) => {
+//       if (validator.isURL(value, { protocols: ['http', 'https', 'ftp'], require_tld: true, require_protocol: true })) {
+//         return value;
+//       }
+//       return helpers.message('Ссылка введена некорректно');
+//     }),
+//     nameRU: Joi.string().required(),
+//     nameEN: Joi.string().required(),
+//     thumbnail: Joi.string().required().custom((value, helpers) => {
+//       if (validator.isURL(value)) {
+//         return value;
+//       }
+//       return helpers.message('Ссылка введена некорректно');
+//     }),
+//     movieId: Joi.number().required(),
+//   }),
+// });
 
 // app.use(routes);
 app.use(router);
