+++ routes/movies.js
@@ -1,7 +1,9 @@
 const movieRouter = require('express').Router();
+// const { validator } = require('validator');
+const { validateMovie } = require('../middlewares/validation');
 const { createMovie, getMovies, deleteMovie } = require('../controllers/movies');
 
-movieRouter.post('/', createMovie);
+movieRouter.post('/', validateMovie, createMovie);
 movieRouter.get('/', getMovies);
 movieRouter.delete('/:movieId', deleteMovie);
